requirements.txt
setuptools
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
numpy
openai-whisper
